moved edit to be subgroup of detail
