+++ src/app/modules/project/details/edit/edit.component.scss
