+++ src/app/modules/project/details/edit/edit.component.ts
@@ -44,6 +44,7 @@ import { QuillUtils } from 'src/app/utils/quill.utils';
 })
 export class EditComponent implements OnInit {
   @Input() project: Project;
+  @Input() projectCategories: ProjectCategory;
   /**
    * Configuration for file-picker
    */
@@ -216,6 +217,7 @@ export class EditComponent implements OnInit {
     this.uploadingFiles = true;
     const editedProject: ProjectUpdate = this.editProjectForm.value;
     editedProject.collaborators = this.collaborators;
+    console.log(this.categories);
     editedProject.categories = this.categories.filter(category => category.selected);
 
     const selectedCallToActions = this.callToActions.callToActionOptions
