+++ src/app/modules/project/project-routing.module.ts
@@ -14,7 +14,7 @@
  *   along with this program, in the LICENSE.md file in the root project directory.
  *   If not, see https://www.gnu.org/licenses/lgpl-3.0.txt
  */
-import { EditComponent } from './edit/edit.component';
+import { EditComponent } from './details/edit/edit.component';
 import { EmbedComponent } from './embed/embed.component';
 import { OverviewComponent } from './overview/overview.component';
