+++ src/app/modules/project/project.module.ts
@@ -19,8 +19,8 @@ import { CollaboratorComponent } from './collaborator/collaborator.component';
 import { BottomDrawerComponent } from './details/bottom-drawer/bottom-drawer.component';
 import { SettingsDropdownComponent } from './details/bottom-drawer/settings-dropdown/settings-dropdown.component';
 import { DetailsComponent } from './details/details.component';
+import { EditComponent } from './details/edit/edit.component';
 import { SummaryComponent } from './details/summary/summary.component';
-import { EditComponent } from './edit/edit.component';
 import { EmbedButtonComponent } from './embed-button/embed-button.component';
 import { EmbedComponent } from './embed/embed.component';
 import { HighlightsModalComponent } from './highlights-modal/highlights-modal.component';
